Reorganise the agents into read and write
